// web/src/api/generated/models/workflow-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 工作流响应
 */
export interface WorkflowResponse {
    /**
     * 工作流ID
     */
    'id'?: number;
    /**
     * 工作流名称
     */
    'name'?: string;
    /**
     * 工作流描述
     */
    'description'?: string;
    /**
     * 工作流定义JSON字符串
     */
    'definition'?: string;
    /**
     * 工作流状态
     */
    'status'?: WorkflowResponseStatusEnum;
    /**
     * 版本号
     */
    'version'?: number;
    /**
     * 创建者ID
     */
    'creatorId'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
    'public'?: boolean;
}

export const WorkflowResponseStatusEnum = {
    DRAFT: 'DRAFT',
    PUBLISHED: 'PUBLISHED',
    ARCHIVED: 'ARCHIVED'
} as const;

export type WorkflowResponseStatusEnum = typeof WorkflowResponseStatusEnum[keyof typeof WorkflowResponseStatusEnum];