// web/src/api/generated/models/update-workflow-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 更新工作流基本信息请求
 */
export interface UpdateWorkflowRequest {
    /**
     * 工作流名称
     */
    'name'?: string;
    /**
     * 工作流描述
     */
    'description'?: string;
}